# src/internal/docs/swagger.yaml
basePath: /
definitions:
...
  models.Err:
    properties:
      description:
        description: the error description for the developer, aka err.Error()
        type: string
      message:
        description: the displayed error text for the user
        type: string
      name:
        type: string
      status:
        type: integer
    type: object
...
  models.Response:
    properties:
      data: {}
      timestamp:
        type: string
    type: object
  models.UserInput:
    properties:
      active:
        type: boolean
      user:
        type: string
    type: object
  models.UserLogin:
    properties:
      authId:
        type: string
      username:
        type: string
    type: object
  models.UserOutput:
    properties:
      _id:
        type: string
      active:
        type: boolean
      createdAt:
        type: string
      username:
        type: string
    type: object
host: 127.0.0.1:3000
info:
  contact:
    email: ??
    name: Tovah Parnes
  description: A server for BirdAI API, for managing users, admin, birds, posts and
    more.
  license:
    name: MIT License
    url: https://opensource.org/license/mit/
  termsOfService: http://swagger.io/terms/
  title: BirdAI API
  version: "2.0"
paths:
...
  /users/:
    post:
      consumes:
      - application/json
      description: Login a user or create a new user if there is no existing user
      parameters:
      - description: user
        in: body
        name: set
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      summary: Login a user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete given user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Delete given user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.UserOutput'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "410":
          description: Gone
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update given user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Update given user
      tags:
      - users
...
  /users/list:
    get:
      consumes:
      - application/json
      description: List all users of a specified set
      parameters:
      - description: Set of users
        in: query
        name: set
        type: integer
      - description: Search parameter for user
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.UserOutput'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: List all users of a specified set
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.UserOutput'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "410":
          description: Gone
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get current user
      tags:
      - users
schemes:
- http
securityDefinitions:
  Bearer:
    description: '>- Enter the token with the `Bearer: ` prefix, e.g. "Bearer eyJhbGciOiJIUzI1NiJ9.e30.DXCaKJxPJq8JhXOZIN4yWdBwJ4sraVbwH9imDCHZPaA".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"